implementacion de todosas las operaciones de manera general para los tres problemas
